added db model and migration

+++ internal/models/models.go
@@ -1,8 +1,8 @@
 package models
 
 type RangeCounter struct {
-	ID       uint `gorm:"primaryKey;column:id"`
-	LastUsed uint `gorm:"not null;column:last_used"`
+	ID       uint  `gorm:"primaryKey;column:id"`
+	LastUsed int64 `gorm:"not null;column:last_used"`
 }
 
 func (RangeCounter) TableName() string {
